Flatten `submit`.
